Update README.md
Exploring Uber Data with Python!

In this project, I used Python and Jupyter notebook to dive into Uber ride data and answer critical questions like:

1.Which ride category is the most booked?

2. Why do people book rides?

3.Peak hours and days for Uber bookings

4.Months with the lowest ride frequency

5. Average ride distance

This hands-on analysis has deepened my understanding of data-driven decision-making in the transportation industry. Excited to share my journey of learning and discovery!

Special thanks to The iScale for guiding this project.